requirements.txt
click==6.7
cssmin==0.2.0
Flask-Assets==0.12
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
jsmin==2.2.2
MarkupSafe==1.0
newspaper3k==0.2.2
nltk==3.2.4
webassets==0.12.1
Werkzeug==0.12.2